functions/recognizer/recognizer__mosaic/requirements.txt
fastapi
uvicorn
opencv-python-headless
pillow
numpy
